Updated dependencies.
Added a flag to the CBL import to use ComicVine parsing for CBL Import (and set to true as all CBLs to my knowledge are using this new library).

Started working on a bug where the validation issue UI code is rendered when it shouldn't. Looks like an angular bug.