Fixed an edge case in LowestSeriesFolder code:
/ <- Lib root assume
/love hina/love hina/v01.cbz
/love hina/specials/sp01.cbz

The lowest series folder was /love hina/love hina/ for some reason, meaning series scan wasn't getting the full series (all folders) scanned in

I'm not sure if I'm still on a good track or not with the scanner loop. It works on some stuff and others not well. But I am finding bugs in other pieces of code, so I will continue on this path to hopefully cleanup more of the scanner.