Update browserslist to support more versions
This updates the .browserslistrc file to not be so strict around what browsers it supports. The old config only had a global browser audience coverage of 23.8 % vs the `defaults` setting of 89.6 %.

This should let mobile browsers load the page much faster based on the testing that was done with a small sample group.