Actually fix tests on Windows
It seems that the MockFileSystem defaults to C:/ as root when none is given. While GitHub actions (apparently) don't run on the C drive luckily neither are my GitHub repo's on my Windows PC.

This also updates the other DateTime tests to be inline with the UTC fix to ensure they're testing the right thing.

Please work now