Overhauled the reading list detail page to be more aligned with series detail page.
Added the ability to see all people across the reading list, the time it would take to read it and age rating.

By default, the list will not be in edit mode and you will need to enable it with the new reorder list button. Remove is also no longer on items when in non-edit mode.

Added a warning in edit mode for when drag and drop is disabled as some users have reported as an issue before, but it is by design.