Change SideNav Drag&Drop to edit mode
There are still some issues with this, the handle isn't an actually
handle and everything works. And the dragpreview isn't very nice.

I tried solving both, but don't have the time for now. I think the issue
has to do with some custom css (I don't understand how the icon and text
are spaced lol, and possibly with cdkDrag and cdkDraghandle not being in
the same file? Not sure