Sort search results by length to ensure sort titles are always present Fixes #3992
Ideally we would give a score on likelyhood, I.e. where does the query
appear, how long is the query alr etc. But SQLite isn't the most
performant so this should work nice enough for now.

These better sortings would be nice to have in a full blown search
if/when we make one as it doesn't need to auto update. I.e. only when
search is pressed